refactor: clean up code review findings
- Replace magic numbers buf[24][32] with TouchBoardWrapper::ROWS/COLS constants
- Remove redundant fps_count_ variable, use frame_times_.size() directly
- Fix usb_pid type mismatch: int32_t → uint16_t in TouchBoardWrapper
- Extract wait_for_new_frame() to deduplicate read_tactile/read_tactile_raw
- Reuse sanitize_sn from hand_zenoh_bridge instead of reimplementing
- Remove dead code (wait method check) in integration_demo.py
- Remove duplicate test (test_tactile_bridge_exists ≡ test_tactile_bridge_init)
- Demote verbose startup logs from INFO to DEBUG in HandBridge

+++ src/touch_board_wrapper.hpp
@@ -18,7 +18,7 @@ public:
     static constexpr int COLS = wujihandcpp::device::TouchBoard::COLS;
 
     explicit TouchBoardWrapper(
-        std::optional<std::string> serial_number, int32_t usb_pid, uint16_t usb_vid)
+        std::optional<std::string> serial_number, uint16_t usb_pid, uint16_t usb_vid)
         : board_(serial_number ? serial_number->c_str() : nullptr, usb_pid, usb_vid) {}
 
     /// Blocking read, returns normalized float32 (24,32). Throws on timeout.
